Modify createLoan to work with a list of device IDs
This allow the frontend to send multiple device IDs at once when
creating a loan. createLoan.dto now accepts an array of device IDs.
The transaction logic in loan.service has it timeout set to 20s in case
the number of devices is large, which can take longer to process. The
service will now return a list of created loans.