Add createTransfer function and fixed business logic errors
The original code for getTransferRequests followed an incorrect logic
flow where fromUserId was the request sender, which should be toUserId
(a transfer goes fromUserId -> toUserId, the one receiving the device
should be the requester). This has been corrected in the updated code,
and createTransfer is coded in this way as well.